Reverts all changes - isolating out only test changes.
