+++ static_frame/core/index_hierarchy.py
@@ -1,88 +1,85 @@
-import functools
-import itertools
 import typing as tp
-from copy import deepcopy
 from ast import literal_eval
+from copy import deepcopy
+from itertools import chain
+from itertools import zip_longest
+from itertools import repeat
 
 import numpy as np
-from arraykit import (
-    name_filter,
-)
-from static_frame.core.container_util import (
-    key_from_container_key,
-    matmul,
-    rehierarch_from_type_blocks,
-    sort_index_for_order,
-)
-from static_frame.core.display import (
-    Display,
-    DisplayActive,
-    DisplayHeader,
-)
-from static_frame.core.util import array_to_duplicated, arrays_equal
+from arraykit import name_filter
+
+from static_frame.core.array_go import ArrayGO
+from static_frame.core.container_util import index_from_optional_constructor
+from static_frame.core.container_util import matmul
+from static_frame.core.container_util import rehierarch_from_type_blocks
+from static_frame.core.container_util import key_from_container_key
+from static_frame.core.container_util import sort_index_for_order
+
+from static_frame.core.display import Display
+from static_frame.core.display import DisplayActive
 from static_frame.core.display_config import DisplayConfig
+from static_frame.core.display import DisplayHeader
 from static_frame.core.doc_str import doc_inject
-from static_frame.core.exception import ErrorInitIndex, ErrorInitIndexNonUnique
+
+from static_frame.core.exception import ErrorInitIndex
 from static_frame.core.hloc import HLoc
-from static_frame.core.index import (
-    ILoc,
-    Index,
-    IndexGO,
-    immutable_index_filter,
-)
+from static_frame.core.index import ILoc
+
+from static_frame.core.index import Index
+from static_frame.core.index import IndexGO
+from static_frame.core.util import iterable_to_array_1d
+from static_frame.core.util import PositionsAllocator
 from static_frame.core.index import mutable_immutable_index_filter
-from static_frame.core.index_auto import RelabelInput
 from static_frame.core.index_base import IndexBase
-from static_frame.core.index_datetime import IndexDatetime
+from static_frame.core.index_level import IndexLevel
+from static_frame.core.index_level import IndexLevelGO
 from static_frame.core.index_level import TreeNodeT
+from static_frame.core.index_auto import RelabelInput
+from static_frame.core.index_datetime import IndexDatetime
+
 from static_frame.core.node_dt import InterfaceDatetime
-from static_frame.core.node_iter import (
-    IterNodeApplyType,
-    IterNodeDepthLevel,
-    IterNodeType,
-)
-from static_frame.core.node_re import InterfaceRe
-from static_frame.core.node_selector import (
-    InterfaceAsType,
-    InterfaceGetItem,
-    TContainer,
-)
+from static_frame.core.node_iter import IterNodeApplyType
+from static_frame.core.node_iter import IterNodeDepthLevel
+from static_frame.core.node_iter import IterNodeType
+from static_frame.core.node_selector import InterfaceAsType
+from static_frame.core.node_selector import InterfaceGetItem
+from static_frame.core.node_selector import TContainer
 from static_frame.core.node_str import InterfaceString
 from static_frame.core.node_transpose import InterfaceTranspose
-from static_frame.core.style_config import StyleConfig
+from static_frame.core.node_re import InterfaceRe
 from static_frame.core.type_blocks import TypeBlocks
-from static_frame.core.util import (
-    CONTINUATION_TOKEN_INACTIVE,
-    DEFAULT_SORT_KIND,
-    DTYPE_BOOL,
-    DTYPE_OBJECT,
-    INT_TYPES,
-    NAME_DEFAULT,
-    NULL_SLICE,
-    BoolOrBools,
-    DepthLevelSpecifier,
-    DtypeSpecifier,
-    GetItemKeyType,
-    IndexConstructor,
-    IndexConstructors,
-    IndexInitializer,
-    NameType,
-    PositionsAllocator,
-    UFunc,
-    array2d_to_array1d,
-    array2d_to_tuples,
-    array_sample,
-    intersect2d,
-    isin,
-    isin_array,
-    isna_array,
-    iterable_to_array_1d,
-    iterable_to_array_2d,
-    key_to_datetime_key,
-    setdiff2d,
-    union2d,
-    ufunc_unique,
-)
+
+from static_frame.core.util import DEFAULT_SORT_KIND
+from static_frame.core.util import DepthLevelSpecifier
+from static_frame.core.util import DtypeSpecifier
+from static_frame.core.util import EMPTY_TUPLE
+from static_frame.core.util import DTYPE_BOOL
+from static_frame.core.util import DTYPE_OBJECT
+from static_frame.core.util import GetItemKeyType
+from static_frame.core.util import IndexConstructor
+from static_frame.core.util import IndexConstructors
+from static_frame.core.util import IndexInitializer
+from static_frame.core.util import INT_TYPES
+from static_frame.core.util import intersect2d
+from static_frame.core.util import isin
+from static_frame.core.util import NAME_DEFAULT
+from static_frame.core.util import NameType
+from static_frame.core.util import NULL_SLICE
+from static_frame.core.util import setdiff2d
+from static_frame.core.util import UFunc
+from static_frame.core.util import union2d
+from static_frame.core.util import array2d_to_array1d
+from static_frame.core.util import array2d_to_tuples
+from static_frame.core.util import iterable_to_array_2d
+from static_frame.core.util import array_sample
+from static_frame.core.util import key_to_datetime_key
+from static_frame.core.util import concat_resolved
+from static_frame.core.util import CONTINUATION_TOKEN_INACTIVE
+from static_frame.core.util import BoolOrBools
+from static_frame.core.util import isna_array
+
+from static_frame.core.style_config import StyleConfig
+
 
 if tp.TYPE_CHECKING:
     from pandas import DataFrame #pylint: disable=W0611 #pragma: no cover
@@ -93,108 +90,29 @@ if tp.TYPE_CHECKING:
 IH = tp.TypeVar('IH', bound='IndexHierarchy')
 
 
-def build_indexers_from_product(lists: tp.List[tp.Sequence[tp.Any]]) -> tp.List[np.ndarray]:
-    """
-    Creates a list of indexer arrays for the product of a list of lists.
-
-    Assumes the lists are unique.
-
-    This is equivalent to: ``np.array(list(itertools.product(*lists)))``
-    except it scales incredibly well.
-
-    It observes that the indexers for a product will look like this:
-
-    Example:
-
-    >>> lists = [[1, 2, 3], [4, 5, 6]]
-    >>> build_indexers_from_product(lists)
-    [
-        array([0, 0, 0, 1, 1, 1, 2, 2, 2]),
-        array([0, 1, 2, 0, 1, 2, 0, 1, 2]),
-    ]
-
-    >>> lists = [[1, 2, 3], [4, 5, 6], [7, 8, 9]]
-    >>> build_indexers_from_product(lists)
-    [
-        array([0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2]),
-        array([0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1]),
-        array([0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2]),
-    ]
-    """
-
-    padded_lengths = np.full(len(lists) + 2, 1, dtype=int)
-    padded_lengths[1:-1] = tuple(map(len, lists))
-
-    all_group_reps = np.cumprod(padded_lengths)[:-2]
-    all_index_reps = np.cumprod(padded_lengths[::-1])[-3::-1]
-
-    result = []
-
-    for i, (group_reps, index_reps) in enumerate(zip(all_group_reps, all_index_reps)):
-        subsection = np.hstack(
-            tuple(
-                map(
-                    # Repeat each index (i.e. element) `index_reps` times
-                    functools.partial(np.tile, reps=index_reps),
-                    range(padded_lengths[i+1]),
-                )
-            )
-        )
-
-        # Repeat each section `index_reps` times
-        indexer = np.tile(subsection, reps=group_reps)
-        indexer.flags.writeable = False
-        result.append(indexer)
-
-    return result
-
-
-def blocks_to_container(blocks: tp.Iterator[np.ndarray]) -> np.ndarray:
-    return TypeBlocks.from_blocks(blocks).values
-
-
-def _mask_to_slice_or_ilocs(mask: np.ndarray) -> tp.Union[slice, np.ndarray, int]:
-    assert mask.dtype == DTYPE_BOOL
-
-    valid_ilocs = PositionsAllocator.get(len(mask))[mask]
-
-    if len(valid_ilocs) == 1:
-        return valid_ilocs[0]
-
-    if len(valid_ilocs) == len(mask):
-        return NULL_SLICE
-
-    steps = ufunc_unique(valid_ilocs[1:] - valid_ilocs[:-1])
-
-    if len(steps) == 1:
-        [step] = steps
-        return slice(valid_ilocs[0], valid_ilocs[-1] + 1, None if step == 1 else step)
-
-    return valid_ilocs
-
-
 #-------------------------------------------------------------------------------
 class IndexHierarchy(IndexBase):
     '''A hierarchy of :obj:`Index` objects, defined as a strict tree of uniform depth across all branches.'''
 
     __slots__ = (
-            '_name',
-            '_indices',
-            '_indexers',
+            '_levels',
             '_blocks',
-            '_values',
             '_recache',
+            '_name',
             )
+    _levels: IndexLevel
+    _blocks: TypeBlocks # should be tp.Optional[TypeBlocks] but many typing changes required
+    _recache: bool
     _name: NameType
-    _indices: tp.List[Index] # Of index objects
-    _indexers: tp.List[np.ndarray] # integer arrays
-    _blocks: TypeBlocks
-    _values: np.ndarray
+
+    # Temporary type overrides, until indices are generic.
+    # __getitem__: tp.Callable[['IndexHierarchy', tp.Hashable], tp.Tuple[tp.Hashable, ...]]
 
     # _IMMUTABLE_CONSTRUCTOR is None from IndexBase
     # _MUTABLE_CONSTRUCTOR will be defined after IndexHierarhcyGO defined
 
     _INDEX_CONSTRUCTOR = Index
+    _LEVEL_CONSTRUCTOR = IndexLevel
 
     _UFUNC_UNION = union2d
     _UFUNC_INTERSECTION = intersect2d
@@ -205,29 +123,10 @@ class IndexHierarchy(IndexBase):
     # constructors
 
     @classmethod
-    def _build_index_constructors(cls, index_constructors: IndexConstructors, depth: int) -> IndexConstructors:
-        if index_constructors is None:
-            return [cls._INDEX_CONSTRUCTOR for _ in range(depth)]
-
-        if callable(index_constructors): # support a single constrctor
-            return [index_constructors for _ in range(depth)]
-
-        return index_constructors
-
-    @classmethod
-    def _build_name_from_indices(cls, indices: tp.List[Index]) -> tp.Optional[tp.Tuple[tp.Hashable, ...]]:
-        # build name from index names, assuming they are all specified
-        name: tp.Tuple[tp.Hashable, ...] = tuple(index.name for index in indices)
-        if any(n is None for n in name):
-            return None
-
-        return name
-
-    @classmethod
     def from_product(cls: tp.Type[IH],
             *levels: IndexInitializer,
             name: NameType = None,
-            index_constructors: IndexConstructors = None,
+            index_constructors: tp.Optional[IndexConstructors] = None,
             ) -> IH:
         '''
         Given groups of iterables, return an ``IndexHierarchy`` made of the product of a values in those groups, where the first group is the top-most hierarchy.
@@ -241,54 +140,65 @@ class IndexHierarchy(IndexBase):
             :obj:`static_frame.IndexHierarchy`
 
         '''
-        if len(levels) == 1:
-            raise ErrorInitIndex('Cannot create IndexHierarchy from only one level.')
-
         indices = [] # store in a list, where index is depth
-
-        index_constructors = cls._build_index_constructors(index_constructors, depth=len(levels))
-
-        for lvl, constructor in itertools.zip_longest(levels, index_constructors):
-            if constructor is None:
-                raise ErrorInitIndex(f'Levels and index_constructors must be the same length.')
-
-            # we call the constructor on all lvl, even if it is already an Index
-            # This will raise if any incoming levels are not unique
-            if isinstance(lvl, Index):
-                indices.append(immutable_index_filter(lvl))
+        if index_constructors is None:
+            for lvl in levels:
+                if not isinstance(lvl, Index): # Index, not IndexBase
+                    lvl = cls._INDEX_CONSTRUCTOR(lvl)
+                indices.append(lvl)
+        else:
+            if callable(index_constructors): # support a single constrctor
+                pair_iter = zip(levels, repeat(index_constructors))
             else:
+                pair_iter = zip_longest(levels, index_constructors)
+
+            for lvl, constructor in pair_iter:
+                if constructor is None:
+                    raise ErrorInitIndex(f'Levels and index_constructors must be the same length.')
+                # we call the constructor on all lvl, even if it is already an Index
                 indices.append(constructor(lvl))
 
+        if len(indices) == 1:
+            raise ErrorInitIndex('Cannot create IndexHierarchy from only one level.')
+
+        # build name from index names, assuming they are all specified
         if name is None:
-            name = cls._build_name_from_indices(indices)
+            name = tuple(index.name for index in indices)
+            if any(n is None for n in name):
+                name = None
 
-        indexers = build_indexers_from_product(indices)
+        targets_previous = None
 
-        return cls(
-            name=name,
-            indices=indices,
-            indexers=indexers,
-        )
+        # need to walk up from bottom to top
+        # get depth pairs and iterate over those
+        depth = len(indices) - 1
+        while depth > 0:
+            index = indices[depth]
+            index_up = indices[depth - 1]
+            # for each label in the next-up index, we need a reference to this index with an offset of that index (or level)
+            targets = np.empty(len(index_up), dtype=object)
 
-    @classmethod
-    def _from_tree(cls, tree: TreeNodeT) -> tp.Iterator[tp.Tuple[tp.Hashable, ...]]:
-        values: tp.List[tp.Tuple[tp.Hashable, ...]] = []
-        for label, subtree in tree.items():
-            if isinstance(subtree, dict):
-                for row in cls._from_tree(subtree):
-                    yield (label, *row)
-            else:
-                for row in subtree: # type: ignore
-                    yield (label, row)
+            offset = 0
+            for idx, _ in enumerate(index_up):
+                # this level does not have targets, only an index (as a leaf)
+                level = cls._LEVEL_CONSTRUCTOR(index=index,
+                        offset=offset,
+                        targets=targets_previous)
 
-        yield from values
+                targets[idx] = level
+                offset += len(level)
+            targets_previous = ArrayGO(targets, own_iterable=True)
+            depth -= 1
+
+        level = cls._LEVEL_CONSTRUCTOR(index=index_up, targets=targets_previous)
+        return cls(level, name=name)
 
     @classmethod
     def from_tree(cls: tp.Type[IH],
             tree: TreeNodeT,
             *,
             name: NameType = None,
-            index_constructors: IndexConstructors = None,
+            index_constructors: tp.Optional[IndexConstructors] = None,
             ) -> IH:
         '''
         Convert into a ``IndexHierarchy`` a dictionary defining keys to either iterables or nested dictionaries of the same.
@@ -296,11 +206,13 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`static_frame.IndexHierarchy`
         '''
-        return cls.from_labels(
-            labels=cls._from_tree(tree),
-            name=name,
-            index_constructors=index_constructors,
-        )
+        return cls(
+                cls._LEVEL_CONSTRUCTOR.from_tree(
+                        tree,
+                        index_constructors
+                        ),
+                name=name,
+                )
 
     @classmethod
     def from_labels(cls: tp.Type[IH],
@@ -308,8 +220,8 @@ class IndexHierarchy(IndexBase):
             *,
             name: NameType = None,
             reorder_for_hierarchy: bool = False,
-            index_constructors: IndexConstructors = None,
-            depth_reference: tp.Optional[DepthLevelSpecifier] = None,
+            index_constructors: tp.Optional[IndexConstructors] = None,
+            depth_reference: tp.Optional[int] = None,
             continuation_token: tp.Union[tp.Hashable, None] = CONTINUATION_TOKEN_INACTIVE
             ) -> IH:
         '''
@@ -319,6 +231,7 @@ class IndexHierarchy(IndexBase):
             labels: an iterator or generator of tuples.
             *,
             name:
+            reorder_for_hierarchy: reorder the labels to produce a hierarchable Index, assuming hierarchability is possible.
             index_constructors:
             depth_reference:
             continuation_token: a Hashable that will be used as a token to identify when a value in a label should use the previously encountered value at the same depth.
@@ -326,20 +239,34 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`static_frame.IndexHierarchy`
         '''
+        if reorder_for_hierarchy:
+            if continuation_token != CONTINUATION_TOKEN_INACTIVE:
+                raise RuntimeError('continuation_token not supported when reorder_for_hiearchy')
+            # use from_records to ensure approprate columnar types
+            from static_frame import Frame
+            index_labels = Frame.from_records(labels)._blocks
+            # this will reorder and create the index using this smae method, passed as cls.from_labels
+            index, _ = rehierarch_from_type_blocks(
+                    labels=index_labels,
+                    depth_map=range(index_labels.shape[1]), # keep order
+                    index_cls=cls,
+                    index_constructors=index_constructors,
+                    name=name,
+                    )
+            return index #type: ignore
+
         labels_iter = iter(labels)
-
         try:
-            label_row = next(labels_iter)
+            first = next(labels_iter)
+            labels_empty = False
         except StopIteration:
-            labels_are_empty = True
-        else:
-            labels_are_empty = False
+            labels_empty = True
 
-        if labels_are_empty:
+        if labels_empty:
+            # if iterable is empty, must discover depth
             if isinstance(labels, np.ndarray) and labels.ndim == 2:
                 # if this is a 2D array, we can get the depth
                 depth = labels.shape[1]
-
                 if depth == 0: # an empty 2D array can have 0 depth
                     pass # do not set depth_reference, assume it is set
                 elif ((depth_reference is None and depth > 1)
@@ -348,70 +275,67 @@ class IndexHierarchy(IndexBase):
                 else:
                     raise ErrorInitIndex(f'depth_reference provided {depth_reference} does not match depth of supplied array {depth}')
 
-            if not isinstance(depth_reference, INT_TYPES):
-                raise ErrorInitIndex('depth_reference must be an integer when labels are empty.')
+            levels = cls._LEVEL_CONSTRUCTOR(
+                    cls._INDEX_CONSTRUCTOR(EMPTY_TUPLE),
+                    depth_reference=depth_reference)
+            return cls(levels=levels, name=name)
 
-            if depth_reference == 1:
-                raise ErrorInitIndex('Cannot create IndexHierarchy from only one level.')
+        depth = len(first)
+        # minimum permitted depth is 2
+        if depth < 2:
+            raise ErrorInitIndex('Cannot create an IndexHierarchy from only one level.')
+        if index_constructors is not None and len(index_constructors) != depth:
+            raise ErrorInitIndex('If providing index constructors, number of index constructors must equal depth of IndexHierarchy.')
 
-            return cls(
-                indices=[cls._INDEX_CONSTRUCTOR(()) for _ in range(depth_reference)],
-                indexers=[PositionsAllocator.get(0) for _ in range(depth_reference)],
-                name=name
-            )
+        depth_max = depth - 1
+        depth_pre_max = depth - 2
 
-        depth = len(label_row)
+        token = object()
+        observed_last = [token for _ in range(depth)]
 
-        if depth == 1:
-            raise ErrorInitIndex('Cannot create IndexHierarchy from only one level.')
+        tree: tp.Any = dict() # order assumed and necessary
+        # put first back in front
+        for label in chain((first,), labels_iter):
+            if len(label) != depth:
+                raise ErrorInitIndex(f'Inconsistent label depth: expected {depth}, got {len(label)}')
 
-        index_constructors = cls._build_index_constructors(index_constructors, depth=depth)
+            current = tree # NOTE: over the life of this loop, current can be a dict or a list
+            # each label is an iterable
+            for d, v in enumerate(label):
+                if continuation_token is not CONTINUATION_TOKEN_INACTIVE:
+                    if v == continuation_token and observed_last[d] is not token:
+                        # only set v to observed_last if it is not the token; if we have not had a previous value that is not a token, the only thing to do is keep v unchanged
+                        v = observed_last[d]
 
-        hash_maps: tp.List[tp.Dict[tp.Hashable, int]] = [{} for _ in range(depth)]
-        indexers: tp.List[tp.List[int]] = [[] for _ in range(depth)]
+                # shared implementation with from_labels -----------------------
+                if d < depth_pre_max:
+                    if v not in current:
+                        current[v] = dict() # order necessary
+                    else: # can only fetch this node (and not create a new node) if this is the sequential predecessor
+                        if v != observed_last[d]:
+                            raise ErrorInitIndex(f'invalid tree-form for IndexHierarchy: {v} in {label} cannot follow {observed_last[d]} when {v} has already been defined.')
+                    current = current[v]
+                    observed_last[d] = v
+                elif d < depth_max:
+                    if v not in current:
+                        current[v] = list()
+                    else: # cannot just fetch this list if it is not the predecessor
+                        if v != observed_last[d]:
+                            raise ErrorInitIndex(f'invalid tree-form for IndexHierarchy: {v} in {label} cannot follow {observed_last[d]} when {v} has already been defined.')
+                    current = current[v]
+                    observed_last[d] = v
+                elif d == depth_max:
+                    # if there are redundancies here they will be caught in index creation
+                    current.append(v)
+                else: # NOTE: cannot get here with length check above
+                    raise ErrorInitIndex('label exceeded expected depth', label) #pragma: no cover
+                # shared implementation with _from_type_blocks -----------------
 
-        prev_row: tp.Sequence[tp.Hashable] = ()
-
-        while True:
-            for i_zip, (hash_map, indexer, val) in enumerate(zip(hash_maps, indexers, label_row)):
-                if val == continuation_token:
-                    if prev_row:
-                        i: int = indexer[-1]
-                        val = prev_row[i_zip]
-                    else:
-                        i = len(hash_map)
-                        hash_map[val] = len(hash_map)
-                elif val not in hash_map:
-                    i = len(hash_map)
-                    hash_map[val] = len(hash_map)
-                else:
-                    i = hash_map[val]
-
-                indexer.append(i)
-
-            prev_row = label_row
-            try:
-                label_row = next(labels_iter)
-            except StopIteration:
-                break
-
-            if len(label_row) != depth:
-                raise ErrorInitIndex("All labels must have the same depth.")
-
-        indices = [
-            constructor(hash_map)
-            for constructor, hash_map
-            in zip(index_constructors, hash_maps)
-        ]
-
-        if name is None:
-            name = cls._build_name_from_indices(indices)
-
-        for i in range(depth):
-            indexers[i] = np.array(indexers[i], dtype=int)
-            indexers[i].flags.writeable = False # type: ignore
-
-        return cls(indices=indices, indexers=indexers, name=name)
+        levels = cls._LEVEL_CONSTRUCTOR.from_tree(
+                tree,
+                index_constructors=index_constructors
+                )
+        return cls(levels=levels, name=name)
 
     @classmethod
     def from_index_items(cls: tp.Type[IH],
@@ -427,58 +351,32 @@ class IndexHierarchy(IndexBase):
             items: iterable of pairs of label, :obj:`Index`.
             index_constructor: Optionally provide index constructor for outermost index.
         '''
-        [depth1_constructor, depth2_constructor] = cls._build_index_constructors(index_constructor, depth=2)
+        labels = []
+        index_levels = []
 
-        depth_1_index = []
-        depth_2_index: tp.Optional[IndexGO] = None
-        indexers_1 = []
-        repeats = []
-
+        offset = 0
         for label, index in items:
-            index: IndexGO = depth2_constructor(index) # type: ignore
+            labels.append(label)
             index = mutable_immutable_index_filter(cls.STATIC, index) #type: ignore
-
-            depth_1_index.append(label)
-
-            if depth_2_index is None:
-                # We will grow this in-place
-                depth_2_index = mutable_immutable_index_filter(False, index) # type: ignore
-                new_indexer = PositionsAllocator.get(len(depth_2_index))
-            else:
-                new_labels = index.difference(depth_2_index) # Retains order!
-
-                if new_labels.size:
-                    depth_2_index.extend(new_labels)
-
-                new_indexer = index.iter_label().apply(depth_2_index._loc_to_iloc)
-
-            indexers_1.append(new_indexer)
-            repeats.append(len(index))
-
-        if not depth_1_index:
-            assert depth_2_index is None
-            return cls(
-                indices=[cls._INDEX_CONSTRUCTOR(()) for _ in range(2)],
-                indexers=[PositionsAllocator.get(0) for _ in range(2)],
-                name=name,
+            index_levels.append(cls._LEVEL_CONSTRUCTOR(
+                    index,
+                    offset=offset,
+                    own_index=True)
             )
+            offset += len(index)
 
-        def _repeat(i_repeat_tuple: tp.Tuple[int, int]) -> np.ndarray:
-            i, repeat = i_repeat_tuple
-            return np.tile(i, reps=repeat)
+        targets = ArrayGO(np.array(index_levels, dtype=object), own_iterable=True)
 
-        indexers = []
-        indexers.append(np.hstack(tuple(map(_repeat, enumerate(repeats)))))
-        indexers.append(np.hstack(indexers_1))
-
-        indexers[0].flags.writeable = False
-        indexers[1].flags.writeable = False
-
-        return cls(
-            indices=[depth1_constructor(depth_1_index), depth_2_index], # type: ignore
-            indexers=indexers,
-            name=name,
-        )
+        index_outer = index_from_optional_constructor(labels,
+                    default_constructor=cls._INDEX_CONSTRUCTOR,
+                    explicit_constructor=index_constructor)
+        levels = cls._LEVEL_CONSTRUCTOR(
+                index=index_outer, #type: ignore
+                targets=targets,
+                own_index=True,
+                depth_reference=2, # depth always 2
+                )
+        return cls(levels, name=name)
 
     @classmethod
     def from_labels_delimited(cls: tp.Type[IH],
@@ -486,7 +384,7 @@ class IndexHierarchy(IndexBase):
             *,
             delimiter: str = ' ',
             name: NameType = None,
-            index_constructors: IndexConstructors = None,
+            index_constructors: tp.Optional[IndexConstructors] = None,
             ) -> IH:
         '''
         Construct an :obj:`IndexHierarchy` from an iterable of labels, where each label is string defining the component labels for all hierarchies using a string delimiter. All components after splitting the string by the delimited will be literal evaled to produce proper types; thus, strings must be quoted.
@@ -497,6 +395,8 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`static_frame.IndexHierarchy`
         '''
+
+
         def to_label(label: str) -> tp.Tuple[tp.Hashable, ...]:
 
             start, stop = None, None
@@ -524,7 +424,9 @@ class IndexHierarchy(IndexBase):
                 )
 
     @classmethod
-    def from_names(cls: tp.Type[IH], names: tp.Iterable[tp.Hashable]) -> IH:
+    def from_names(cls: tp.Type[IH],
+            names: tp.Iterable[tp.Hashable]
+            ) -> IH:
         '''
         Construct a zero-length :obj:`IndexHierarchy` from an iterable of ``names``, where the length of ``names`` defines the zero-length depth.
 
@@ -532,21 +434,17 @@ class IndexHierarchy(IndexBase):
             names: Iterable of hashable names per depth.
         '''
         name = tuple(names)
-        if len(name) == 0:
-            raise ErrorInitIndex("names must be non-empty.")
+        depth = len(name)
+        levels = cls._LEVEL_CONSTRUCTOR.from_depth(depth)
+        return cls(levels, name=name)
 
-        return cls(
-            indices=[cls._INDEX_CONSTRUCTOR((), name=name) for name in names],
-            indexers=[PositionsAllocator.get(0) for _ in names],
-            name=name,
-        )
 
     @classmethod
     def _from_type_blocks(cls: tp.Type[IH],
             blocks: TypeBlocks,
             *,
             name: NameType = None,
-            index_constructors: IndexConstructors = None,
+            index_constructors: tp.Optional[IndexConstructors] = None,
             own_blocks: bool = False,
             ) -> IH:
         '''
@@ -558,129 +456,129 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`IndexHierarchy`
         '''
-        indices: tp.List[Index] = []
-        indexers: tp.List[np.ndarray] = []
 
-        if index_constructors is None:
-            constructor_iter = (cls._INDEX_CONSTRUCTOR for _ in range(blocks.shape[1]))
-        elif callable(index_constructors):
-            constructor_iter = (index_constructors for _ in range(blocks.shape[1]))
-        else:
-            constructor_iter = index_constructors # type: ignore
+        depth = blocks.shape[1]
 
-        for i, (block, constructor) in enumerate(itertools.zip_longest(blocks, constructor_iter)):
+        # minimum permitted depth is 2
+        if depth < 2:
+            raise ErrorInitIndex('cannot create an IndexHierarchy from only one level.')
+        if index_constructors is not None and len(index_constructors) != depth:
+            raise ErrorInitIndex('if providing index constructors, number of index constructors must equal depth of IndexHierarchy.')
 
-            if block is None or constructor is None:
-                raise ErrorInitIndex(f'Levels and index_constructors must be the same length.')
+        depth_max = depth - 1
+        depth_pre_max = depth - 2
 
-            unique_values, indexer = ufunc_unique(block.values, return_inverse=True)
+        token = object()
+        observed_last = [token for _ in range(depth)]
 
-            # we call the constructor on all lvl, even if it is already an Index
-            try:
-                indices.append(constructor(unique_values))
-            except ValueError:
-                raise ErrorInitIndex(f'Could construct {constructor.__name__} with values at depth {i}')
+        tree: tp.Any = dict() # order assumed and necessary
 
-            indexer.flags.writeable = False
-            indexers.append(indexer)
+        idx_row_last = -1
+        for (idx_row, d), v in blocks.element_items():
+            if idx_row_last != idx_row:
+                # for each row, we re-set current to the outermost reference
+                current = tree
+                idx_row_last = idx_row
+
+            # shared implementation with from_labels ---------------------------
+            if d < depth_pre_max:
+                if v not in current:
+                    current[v] = dict() # order necessary
+                else: # can only fetch this node (and not create a new node) if this is the sequential predecessor
+                    if v != observed_last[d]:
+                        raise ErrorInitIndex(f'invalid tree-form for IndexHierarchy: {v} cannot follow {observed_last[d]} when {v} has already been defined.')
+                current = current[v]
+                observed_last[d] = v
+            elif d < depth_max: # premax means inner values are a list
+                if v not in current:
+                    current[v] = list()
+                else: # cannot just fetch this list if it is not the predecessor
+                    if v != observed_last[d]:
+                        raise ErrorInitIndex(f'invalid tree-form for IndexHierarchy: {v} cannot follow {observed_last[d]} when {v} has already been defined.')
+                current = current[v]
+                observed_last[d] = v
+            elif d == depth_max:
+                # if there are redundancies here they will be caught in index creation
+                current.append(v)
+            else:
+                # cannot happen with TypeBlocks
+                raise ErrorInitIndex('label exceeded expected depth', v) #pragma: no cover
+            # shared implementation with from_labels ---------------------------
+
+        levels = cls._LEVEL_CONSTRUCTOR.from_tree(
+                tree,
+                index_constructors=index_constructors,
+                depth_reference=depth,
+                )
 
         if index_constructors is not None:
             # If defined, we may have changed columnar dtypes in IndexLevels, and cannot reuse blocks
-            if tuple(blocks.dtypes) != tuple(index.dtype for index in indices):
+            if tuple(blocks.dtypes) != tuple(levels.dtype_per_depth()):
                 blocks = None #type: ignore
                 own_blocks = False
 
-        return cls(
-            indices=indices,
-            indexers=indexers,
-            name=name,
-            _blocks=blocks,
-            _own_blocks=own_blocks,
-        )
+        return cls(levels=levels, name=name, blocks=blocks, own_blocks=own_blocks)
 
     # NOTE: could have a _from_fields (or similar) that takes a sequence of column iterables/arrays
 
-    @staticmethod
-    def _ensure_uniqueness(values: np.ndarray) -> None:
-        duplicates = array_to_duplicated(values, exclude_first=True, exclude_last=False)
-        if any(duplicates):
-            first_duplicate = values[duplicates][0]
-            msg = f'Labels have {sum(duplicates)} non-unique values, including {tuple(first_duplicate)}.'
-            raise ErrorInitIndexNonUnique(msg)
-
-    def _gen_blocks_from_self(self) -> TypeBlocks:
-        def gen_blocks() -> tp.Iterator[np.ndarray]:
-            for i, index in enumerate(self._indices):
-                indexer = self._indexers[i]
-                yield index.values[indexer]
-
-        return TypeBlocks.from_blocks(gen_blocks())
-
     #---------------------------------------------------------------------------
     def __init__(self,
-            indices: tp.Union["IndexHierarchy", tp.List[Index]],
+            levels: tp.Union[IndexLevel, 'IndexHierarchy'],
             *,
-            indexers: tp.List[np.ndarray] = (), # type: ignore
             name: NameType = NAME_DEFAULT,
-            _blocks: tp.Optional[TypeBlocks] = None,
-            _own_blocks: bool = False,
+            blocks: tp.Optional[TypeBlocks] = None,
+            own_blocks: bool = False,
             ):
         '''
         Initializer.
 
         Args:
-            indices: list of :obj:`Index` objects
-            indexers: list of indexer arrays
-            name: name of the IndexHierarchy
+            levels: :obj:`IndexLevels` instance, or, optionally, an :obj`IndexHierarchy` to be used to construct a new :obj`IndexHierarchy`.
         '''
-        # TODO: Really ugly hack. Better to create specialized constructor
-        if isinstance(indices, IndexHierarchy):
-            if indexers:
-                raise ErrorInitIndex('indexers must not be provided when copying an IndexHierarchy')
-            if _blocks is not None:
-                raise ErrorInitIndex('_blocks must not be provided when copying an IndexHierarchy')
 
-            self._indices = [mutable_immutable_index_filter(self.STATIC, idx) for idx in indices._indices]
-            self._indexers = indices._indexers
-            self._name = indices._name
-            self._blocks = indices._blocks
-            return
+        self._blocks = None #type: ignore
 
-        if not all(isinstance(arr, np.ndarray) for arr in indexers):
-            raise ErrorInitIndex("indexers must be numpy arrays.")
+        if isinstance(levels, IndexHierarchy):
+            if not blocks is None:
+                raise ErrorInitIndex('cannot provide blocks when initializing with IndexHierarchy')
+            index_level = levels._levels
+            # if cache is updated, can get blocks
+            if not levels._recache:
+                self._blocks = levels._blocks.copy()
+            # transfer name if not given as arg
+            if name is NAME_DEFAULT:
+                name = levels.name
 
-        if not all(not arr.flags.writeable for arr in indexers):
-            raise ErrorInitIndex("indexers must be read-only.")
+        elif isinstance(levels, IndexLevel):
+            index_level = levels
+            if not blocks is None:
+                self._blocks = blocks if own_blocks else blocks.copy()
 
-        if not all(isinstance(index, Index) for index in indices):
-            raise ErrorInitIndex("indices must be Index's!")
-
-        self._indices = [mutable_immutable_index_filter(self.STATIC, idx) for idx in indices]
-        self._indexers = indexers
-        self._name = None if name is NAME_DEFAULT else name_filter(name)
-
-        if _blocks is not None:
-            if _own_blocks:
-                self._blocks = _blocks
-            else:
-                self._blocks = _blocks.copy()
         else:
-            self._blocks = self._gen_blocks_from_self()
+            raise NotImplementedError(f'no handling for creation from {levels}')
 
-        self._values = self._blocks.values
+        if self.STATIC and index_level.STATIC:
+            self._levels = index_level
+        else: # must deepcopy IndexLevels if not IndexHierarchy not static
+            self._levels = index_level.to_index_level(
+                    cls=self._LEVEL_CONSTRUCTOR
+                    )
 
-        self._ensure_uniqueness(self._values)
-        self._recache = False
+        if self._levels.depth <= 1:
+            raise ErrorInitIndex(f'invalid depth ({self._levels.depth}) for IndexLevels composed in IndexHierarchy')
 
-    def _update_array_cache(self) -> None:
-        pass
+        self._recache = self._blocks is None
+        self._name = None if name is NAME_DEFAULT else name_filter(name)
 
     #---------------------------------------------------------------------------
     def __deepcopy__(self: IH, memo: tp.Dict[int, tp.Any]) -> IH:
+        if self._recache:
+            self._update_array_cache()
+
         obj = self.__new__(self.__class__)
-        obj._indices = deepcopy(self._indices, memo)
-        obj._indexers = deepcopy(self._indexers, memo)
+        obj._levels = deepcopy(self._levels, memo)
         obj._blocks = deepcopy(self._blocks, memo)
+        obj._recache = False
         obj._name = self._name # should be hashable/immutable
 
         memo[id(self)] = obj
@@ -690,13 +588,15 @@ class IndexHierarchy(IndexBase):
         '''
         Return a shallow copy of this IndexHierarchy.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         blocks = self._blocks.copy()
         return self.__class__(
-                indices=self._indices,
-                indexers=self._indexers,
+                levels=self._levels, # immutable, can re-use
                 name=self._name,
-                _blocks=blocks,
-                _own_blocks=True
+                blocks=blocks,
+                own_blocks=True
                 )
 
     def copy(self: IH) -> IH:
@@ -710,33 +610,23 @@ class IndexHierarchy(IndexBase):
 
     def rename(self: IH, name: NameType) -> IH:
         '''
-        Return a new IndexHierarchy with an updated name attribute.
+        Return a new Frame with an updated name attribute.
         '''
-        if self.STATIC:
-            indices = self._indices
-            blocks = self._blocks
-        else:
-            indices = [idx.copy() for idx in self._indices]
-            blocks = self._blocks.copy()
-
-        return self.__class__(
-                indices=indices,
-                indexers=list(self._indexers),
-                name=name,
-                _blocks=blocks,
-                _own_blocks=True,
-                )
+        # do not need to recache
+        # let the constructor handle reuse
+        return self.__class__(self, name=name)
 
     #---------------------------------------------------------------------------
     # interfaces
 
     @property
-    def loc(self) -> InterfaceGetItem["IndexHierarchy"]:
+    def loc(self) -> InterfaceGetItem['IndexHierarchy']:
         return InterfaceGetItem(self._extract_loc) #type: ignore
 
     @property
-    def iloc(self) -> InterfaceGetItem["IndexHierarchy"]:
+    def iloc(self) -> InterfaceGetItem['IndexHierarchy']:
         return InterfaceGetItem(self._extract_iloc) #type: ignore
+
 
     def _iter_label(self,
             depth_level: tp.Optional[DepthLevelSpecifier] = None,
@@ -745,12 +635,21 @@ class IndexHierarchy(IndexBase):
         if depth_level is None: # default to full labels
             depth_level = list(range(self.depth))
 
-        if isinstance(depth_level, INT_TYPES):
-            yield from self._blocks._extract_array(column_key=depth_level)
+        # if no type blocks, use a levels
+        if self._recache:
+            if isinstance(depth_level, INT_TYPES):
+                yield from self._levels.labels_at_depth(depth_level=depth_level)
+            else:
+                yield from zip(
+                        *(self._levels.labels_at_depth(depth_level=d) for d in depth_level)
+                        )
         else:
-            yield from array2d_to_tuples(
-                    self._blocks._extract_array(column_key=depth_level)
-                    )
+            if isinstance(depth_level, INT_TYPES):
+                yield from self._blocks._extract_array(column_key=depth_level)
+            else:
+                yield from array2d_to_tuples(
+                        self._blocks._extract_array(column_key=depth_level)
+                        )
 
     def _iter_label_items(self,
             depth_level: tp.Optional[DepthLevelSpecifier] = None,
@@ -758,6 +657,7 @@ class IndexHierarchy(IndexBase):
         '''This function is not directly called in iter_label or related routines, fulfills the expectations of the IterNodeDepthLevel interface.
         '''
         yield from enumerate(self._iter_label(depth_level=depth_level))
+
 
     @property
     def iter_label(self) -> IterNodeDepthLevel[tp.Any]:
@@ -788,6 +688,12 @@ class IndexHierarchy(IndexBase):
         '''
         Interface for applying string methods to elements in this container.
         '''
+        if self._recache:
+            self._update_array_cache()
+
+        def blocks_to_container(blocks: tp.Iterator[np.ndarray]) -> np.ndarray:
+            return TypeBlocks.from_blocks(blocks).values
+
         return InterfaceString(
                 blocks=self._blocks._blocks,
                 blocks_to_container=blocks_to_container,
@@ -798,17 +704,25 @@ class IndexHierarchy(IndexBase):
         '''
         Interface for applying datetime properties and methods to elements in this container.
         '''
+        if self._recache:
+            self._update_array_cache()
+
+        def blocks_to_container(blocks: tp.Iterator[np.ndarray]) -> np.ndarray:
+            return TypeBlocks.from_blocks(blocks).values
+
         return InterfaceDatetime(
                 blocks=self._blocks._blocks,
                 blocks_to_container=blocks_to_container,
                 )
 
     @property
-    def via_T(self) -> InterfaceTranspose["IndexHierarchy"]:
+    def via_T(self) -> InterfaceTranspose['IndexHierarchy']:
         '''
         Interface for using binary operators with one-dimensional sequences, where the opperand is applied column-wise.
         '''
-        return InterfaceTranspose(container=self)
+        return InterfaceTranspose(
+                container=self,
+                )
 
     def via_re(self,
             pattern: str,
@@ -817,6 +731,12 @@ class IndexHierarchy(IndexBase):
         '''
         Interface for applying regular expressions to elements in this container.
         '''
+        if self._recache:
+            self._update_array_cache()
+
+        def blocks_to_container(blocks: tp.Iterator[np.ndarray]) -> np.ndarray:
+            return TypeBlocks.from_blocks(blocks).values
+
         return InterfaceRe(
                 blocks=self._blocks._blocks,
                 blocks_to_container=blocks_to_container,
@@ -826,11 +746,20 @@ class IndexHierarchy(IndexBase):
 
     #---------------------------------------------------------------------------
 
+    def _update_array_cache(self) -> None:
+        self._blocks = self._levels.to_type_blocks()
+        self._recache = False
+
+    #---------------------------------------------------------------------------
+
     @property # type: ignore
     @doc_inject()
     def mloc(self) -> int:
         '''{doc_int}
         '''
+        if self._recache:
+            self._update_array_cache()
+
         return self._blocks.mloc #type: ignore
 
     @property
@@ -843,8 +772,14 @@ class IndexHierarchy(IndexBase):
         '''
         from static_frame.core.series import Series
 
+        if self._recache:
+            # might use self._levels.dtype_per_depth
+            self._update_array_cache()
+
+        labels: NameType
+
         if self._name and isinstance(self._name, tuple) and len(self._name) == self.depth:
-            labels: NameType = self._name
+            labels = self._name
         else:
             labels = None
 
@@ -858,6 +793,8 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`tp.Tuple[int]`
         '''
+        if self._recache:
+            return self._levels.__len__(), self._levels.depth
         return self._blocks._shape
 
     @property
@@ -878,6 +815,8 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`int`
         '''
+        if self._recache:
+            return self._levels.__len__() * self._levels.depth
         return self._blocks.size
 
     @property
@@ -888,11 +827,16 @@ class IndexHierarchy(IndexBase):
         Returns:
             :obj:`int`
         '''
+        if self._recache:
+            self._update_array_cache()
         return self._blocks.nbytes
 
     #---------------------------------------------------------------------------
 
     def __len__(self) -> int:
+        if self._recache:
+            # avoid full recache
+            return self._levels.__len__()
         return self._blocks.__len__()
 
     @doc_inject()
@@ -907,6 +851,9 @@ class IndexHierarchy(IndexBase):
             {config}
         '''
         config = config or DisplayActive.get()
+
+        if self._recache:
+            self._update_array_cache()
 
         sub_display: tp.Optional[Display] = None
 
@@ -960,6 +907,9 @@ class IndexHierarchy(IndexBase):
                 # we will no longer have type associations per depth
                 return self.__class__.from_labels((), depth_reference=self.depth)
 
+        if self._recache:
+            self._update_array_cache()
+
         if isinstance(other, np.ndarray):
             operand = other
             assume_unique = False
@@ -988,7 +938,9 @@ class IndexHierarchy(IndexBase):
         if both_sized and isinstance(other, IndexHierarchy):
             index_constructors = []
             # depth, and length of index_types, must be equal
-            for cls_self, cls_other in zip(self._index_constructors, other._index_constructors):
+            for cls_self, cls_other in zip(
+                    self._levels.index_types(),
+                    other._levels.index_types()):
                 if cls_self == cls_other:
                     index_constructors.append(cls_self)
                 else:
@@ -1001,19 +953,19 @@ class IndexHierarchy(IndexBase):
                 index_constructors=index_constructors,
                 depth_reference=self.depth)
 
+
     #---------------------------------------------------------------------------
-
-    @property
-    def _index_constructors(self) -> tp.Tuple[tp.Type[Index], ...]:
-        return tuple(index.__class__ for index in self._indices)
-
     def _drop_iloc(self, key: GetItemKeyType) -> 'IndexHierarchy':
         '''Create a new index after removing the values specified by the loc key.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         blocks = TypeBlocks.from_blocks(self._blocks._drop_blocks(row_key=key))
+        index_constructors = tuple(self._levels.index_types())
 
         return self.__class__._from_type_blocks(blocks,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 name=self._name,
                 own_blocks=True
                 )
@@ -1031,6 +983,8 @@ class IndexHierarchy(IndexBase):
         '''
         {}
         '''
+        if self._recache:
+            self._update_array_cache()
         return self._blocks.values
 
     @property
@@ -1041,7 +995,10 @@ class IndexHierarchy(IndexBase):
 
     @property
     def depth(self) -> int: #type: ignore
-        return len(self._indices)
+        if self._recache:
+            # avoid full recache to get depth
+            return self._levels.depth
+        return self._blocks.shape[1]
 
     def values_at_depth(self,
             depth_level: DepthLevelSpecifier = 0
@@ -1052,57 +1009,27 @@ class IndexHierarchy(IndexBase):
         Args:
             depth_level: a single depth level, or iterable depth of depth levels.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         sel: GetItemKeyType
 
         if isinstance(depth_level, INT_TYPES):
             sel = depth_level
         else:
             sel = list(depth_level)
-
         return self._blocks._extract_array(column_key=sel)
 
-    # TODO: VERY SLOW!
     @doc_inject()
     def label_widths_at_depth(self,
             depth_level: DepthLevelSpecifier = 0
             ) -> tp.Iterator[tp.Tuple[tp.Hashable, int]]:
         '''{}'''
-        pos: tp.Optional[int] = None
-
-        if depth_level is None:
-            raise NotImplementedError('depth_level of None is not supported')
-        elif not isinstance(depth_level, INT_TYPES):
-            sel = list(depth_level)
-            if len(sel) == 1:
-                pos = sel.pop()
-        else: # is an int
-            pos = depth_level
-
-        if pos is None:
-            raise NotImplementedError("selecting multiple depth levels is not yet implemented")
-
-        def _extractor(arr: np.ndarray, pos: int) -> tp.Iterator[tp.Tuple[tp.Hashable, int]]:
-            unique, widths = ufunc_unique(arr, return_counts=True)
-            labels = self._indices[pos].values[unique]
-            yield from zip(labels, widths)
-
-        # i.e. depth_level is an int
-        if pos == 0:
-            arr = self._indexers[pos]
-            yield from _extractor(arr, pos)
-            return
-
-        def gen()-> tp.Iterator[tp.Tuple[tp.Hashable, int]]:
-            for outer_level_idxs in itertools.product(*map(range, map(len, self._indices[:pos]))):
-                screen = np.full(self.__len__(), True, dtype=bool)
-
-                for i, outer_level_idx in enumerate(outer_level_idxs):
-                    screen &= (self._indexers[i] == outer_level_idx)
-
-                arr = self._indexers[pos][screen] # type: ignore
-                yield from _extractor(arr, pos)
-
-        yield from gen()
+        if isinstance(depth_level, INT_TYPES):
+            sel = depth_level
+        else:
+            raise NotImplementedError('selection from iterables is not implemented')
+        yield from self._levels.label_widths_at_depth(depth_level=depth_level)
 
     @property
     def index_types(self) -> 'Series':
@@ -1122,13 +1049,18 @@ class IndexHierarchy(IndexBase):
             labels = None
 
         # NOTE: consider caching index_types
-        return Series(self._index_constructors, index=labels, dtype=DTYPE_OBJECT)
+        return Series(self._levels.index_types(), index=labels, dtype=DTYPE_OBJECT)
 
     #---------------------------------------------------------------------------
     def relabel(self, mapper: RelabelInput) -> 'IndexHierarchy':
         '''
         Return a new IndexHierarchy with labels replaced by the callable or mapping; order will be retained. If a mapping is used, the mapping should map tuple representation of labels, and need not map all origin keys.
         '''
+        if self._recache:
+            self._update_array_cache()
+
+        index_constructors = tuple(self._levels.index_types())
+
         if not callable(mapper):
             # if a mapper, it must support both __getitem__ and __contains__
             getitem = getattr(mapper, 'get')
@@ -1141,14 +1073,15 @@ class IndexHierarchy(IndexBase):
 
             return self.__class__.from_labels(gen(),
                     name=self._name,
-                    index_constructors=self._index_constructors,
+                    index_constructors=index_constructors,
                     )
 
         return self.__class__.from_labels(
                 (mapper(x) for x in self._blocks.axis_values(axis=1)), #type: ignore
                 name=self._name,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 )
+
 
     def relabel_at_depth(self,
             mapper: RelabelInput,
@@ -1173,9 +1106,12 @@ class IndexHierarchy(IndexBase):
 
         albeit more efficient.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         if isinstance(depth_level, INT_TYPES):
             depth_level = [depth_level]
-            target_depths = set(depth_level)
+            target_depths: tp.Container[int] = depth_level
         else:
             depth_level = sorted(depth_level)
             target_depths = set(depth_level)
@@ -1189,64 +1125,30 @@ class IndexHierarchy(IndexBase):
         if any(level < 0 or level >= self.depth for level in depth_level):
             raise ValueError(f'Invalid depth level found. Valid levels: [0-{self.depth - 1}]')
 
-        is_callable = callable(mapper)
-
-        # Special handling for full replacements
-        if not is_callable and not hasattr(mapper, 'get'):
-            values, _ = iterable_to_array_1d(mapper, count=len(self))
-
-            if len(values) != len(self):
-                raise ValueError('Iterable must provide a value for each label')
-
-            def gen() -> tp.Iterator[np.ndarray]:
-                for depth_idx in range(self.depth):
-                    if depth_idx in target_depths:
-                        yield values
-                    else:
-                        yield self._blocks._extract_array_column(depth_idx)
-
-            return self.__class__._from_type_blocks(
-                    TypeBlocks.from_blocks(gen()),
+        if callable(mapper) or hasattr(mapper, 'get'):
+            return self.__class__(
+                    levels=self._levels._relabel_at_depth(mapper, depth_level),
                     name=self._name,
-                    index_constructors=self._index_constructors,
-                    own_blocks=True
-                )
+                    )
 
-        mapper_func = mapper if is_callable else mapper.__getitem__ # type: ignore
+        values, _ = iterable_to_array_1d(mapper, count=len(self))
 
-        def get_new_label(label: tp.Hashable) -> tp.Hashable:
-            if is_callable or label in mapper: # type: ignore
-                return mapper_func(label) # type: ignore
-            return label
+        if len(values) != len(self):
+            raise ValueError('Iterable must provide a value for each label')
 
-        new_indices = list(self._indices)
-        new_indexers = list(self._indexers)
-
-        for level in depth_level:
-            index = self._indices[level]
-
-            new_index: tp.Dict[tp.Hashable, int] = {}
-            index_remap: tp.Dict[tp.Hashable, tp.Hashable] = {}
-
-            for label_idx, label in enumerate(index.values):
-                new_label = get_new_label(label)
-
-                if new_label not in new_index:
-                    new_index[new_label] = len(new_index)
+        def gen() -> tp.Iterator[np.ndarray]:
+            for depth_idx in range(self.depth):
+                if depth_idx in target_depths:
+                    yield values
                 else:
-                    index_remap[label_idx] = new_index[new_label]
+                    yield self._blocks._extract_array_column(depth_idx)
 
-            new_indices[level] = index.__class__(new_index)
-
-            indexer = np.array([index_remap.get(i, i) for i in self._indexers[level]])
-            indexer.flags.writeable = False
-            new_indexers[level] = indexer
-
-        return self.__class__(
-            indices=new_indices,
-            indexers=new_indexers,
-            name=self._name,
-        )
+        return self.__class__._from_type_blocks(
+                TypeBlocks.from_blocks(gen()),
+                name=self._name,
+                index_constructors=tuple(self._levels.index_types()),
+                own_blocks=True
+                )
 
     def rehierarch(self: IH,
             depth_map: tp.Sequence[int]
@@ -1254,176 +1156,48 @@ class IndexHierarchy(IndexBase):
         '''
         Return a new :obj:`IndexHierarchy` that conforms to the new depth assignments given be `depth_map`.
         '''
+        if self._recache:
+            self._update_array_cache()
+
+        index_constructors = tuple(self._levels.index_types())
+
         index, _ = rehierarch_from_type_blocks(
                 labels=self._blocks,
                 index_cls=self.__class__,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 depth_map=depth_map,
                 )
         return index #type: ignore
 
     #---------------------------------------------------------------------------
 
-    def _process_key_at_depth(self, depth: int, key: GetItemKeyType) -> tp.Union[slice, np.ndarray]:
-        if depth >= self.depth:
-            raise RuntimeError(f'Invalid depth level for key={key} depth={depth}')
-
-        key_at_depth = key[depth] # type: ignore
-
-        # Key is already a mask!
-        if isinstance(key_at_depth, np.ndarray) and key_at_depth.dtype == DTYPE_BOOL:
-            return key_at_depth
-
-        index_at_depth = self._indices[depth]
-        indexer_at_depth = self._indexers[depth]
-
-        if isinstance(key_at_depth, slice):
-            if key_at_depth.start is not None:
-                start: tp.Optional[int] = index_at_depth.loc_to_iloc(key_at_depth.start) # type: ignore
-            else:
-                start = None
-
-            if key_at_depth.step is not None and not isinstance(key_at_depth.step, INT_TYPES):
-                raise NotImplementedError(f"step must be an integer. What does this even mean? {key_at_depth}")
-
-            if key_at_depth.stop is not None:
-                stop: tp.Optional[int] = index_at_depth.loc_to_iloc(key_at_depth.stop) + 1 # type: ignore
-            else:
-                stop = None
-
-            return isin_array(
-                    array=indexer_at_depth,
-                    array_is_unique=False,
-                    other=np.arange(start, stop, key_at_depth.step),
-                    other_is_unique=True
-                    )
-
-        key_iloc = index_at_depth.loc_to_iloc(key_at_depth)
-
-        if hasattr(key_iloc, "__len__"):
-            if isinstance(key, np.ndarray):
-                return isin_array(
-                        array=indexer_at_depth,
-                        array_is_unique=False,
-                        other=key_iloc,
-                        other_is_unique=True
-                        )
-            return isin(indexer_at_depth, key_iloc)
-
-        return indexer_at_depth == key_iloc
-
-    def _loc_to_iloc(self, key: tp.Union[GetItemKeyType, HLoc]) -> GetItemKeyType:
+    def _loc_to_iloc(self,
+            key: tp.Union[GetItemKeyType, HLoc]
+            ) -> GetItemKeyType:
         '''
-        Given iterable (or instance) of GetItemKeyType, determine the equivalent iloc key.
-
-        When possible, prefer slices.
+        Given iterable of GetItemKeyTypes, apply to each level of levels.
         '''
-        if isinstance(key, ILoc):
+        from static_frame.core.series import Series
+
+        if key.__class__ is ILoc:
             return key.key
 
         if isinstance(key, IndexHierarchy):
+            # default iteration of IH is as tuple
+            return [self._levels.leaf_loc_to_iloc(k) for k in key]
 
-            if not key.depth == self.depth:
-                raise KeyError(f"Key must have the same depth as the index. {key}")
-
-            # TODO: Explore optimizations here
-            return [self._loc_to_iloc(label) for label in key.iter_label()]
-
-        if isinstance(key, np.ndarray) and key.dtype == DTYPE_BOOL:
-            # TODO: Can I just return key?
+        if (key.__class__ is np.ndarray) and key.dtype == DTYPE_BOOL:
             return self.positions[key]
-
-        if isinstance(key, slice):
-            if key == NULL_SLICE:
-                return key
-
-            # reuse - slice_to_inclusive_slice and/or LocMap.map_slice_args
-            if key.start is not None:
-                start: tp.Optional[int] = self._loc_to_iloc(key.start) # type: ignore
-            else:
-                start = None
-
-            if key.step is not None and not isinstance(key.step, INT_TYPES):
-                raise ValueError(f"step must be an integer. What does this even mean? {key}")
-
-            if key.stop is not None:
-                stop: tp.Optional[int] = self._loc_to_iloc(key.stop) + 1 # type: ignore
-            else:
-                stop = None
-
-            return slice(start, stop, key.step)
-
-        if isinstance(key, list):
-            return [self._loc_to_iloc(k) for k in key]
-
-        if isinstance(key, np.ndarray) and key.dtype != DTYPE_BOOL and key.ndim == 2:
-            return [self._loc_to_iloc(k) for k in key]
 
         if key.__class__ is HLoc:
             # unpack any Series, Index, or ILoc into the context of this IndexHierarchy
-            key = tuple(
-                    HLoc(
-                        tuple(
-                            key_from_container_key(self, k, expand_iloc=True)
-                            for k in key # type: ignore
-                        )
-                    )
-                )
+            key = HLoc(tuple(
+                    key_from_container_key(self, k, expand_iloc=True)
+                    for k in key))
         else:
-            # If the key is a series, key_from_container_key will invoke IndexCorrespondence
-            # logic that eventually calls _loc_to_iloc on all the indices of that series.
             key = key_from_container_key(self, key)
-            if isinstance(key, np.ndarray) and key.dtype == DTYPE_BOOL:
-                return PositionsAllocator.get(len(key))[key]
 
-            key = tuple(key)
-
-        if any(isinstance(k, tuple) for k in key):
-            return [self._loc_to_iloc(k) for k in key]
-
-        meaningful_selections = {depth: not (isinstance(k, slice) and k == NULL_SLICE) for depth, k in enumerate(key)}
-
-        can_return_element = all(
-                meaningful and (isinstance(key[depth], str) or not hasattr(key[depth], "__len__"))
-                for depth, meaningful
-                in meaningful_selections.items()
-                )
-
-        meaningful_depths = sum(meaningful_selections.values())
-
-        # Return a slice wherever possible
-        if meaningful_depths == 1:
-
-            depth = next(i for i, meaningful in meaningful_selections.items() if meaningful)
-            mask = self._process_key_at_depth(depth=depth, key=key)
-
-            if isinstance(mask, slice):
-                return mask # Not actually a mask
-
-        else:
-            mask_2d = np.full(self.shape, True, dtype=bool)
-
-            for depth, meaningful in meaningful_selections.items():
-                if not meaningful:
-                    continue
-
-                result = self._process_key_at_depth(depth=depth, key=key)
-
-                if isinstance(result, slice):
-                    mask_2d[:, depth] = False
-                    mask_2d[result, depth] = True
-                else:
-                    mask_2d[:, depth] = result
-
-            mask = mask_2d.all(axis=1)
-            del mask_2d
-
-        result = PositionsAllocator.get(len(mask))[mask]
-
-        # Even if there was one result, unless the HLoc specified all levels, we need to return a list
-        if len(result) == 1 and can_return_element:
-            return result[0]
-        return result
+        return self._levels.loc_to_iloc(key)
 
     def loc_to_iloc(self,
             key: tp.Union[GetItemKeyType, HLoc]
@@ -1441,17 +1215,21 @@ class IndexHierarchy(IndexBase):
             ) -> tp.Union['IndexHierarchy', tp.Tuple[tp.Hashable]]:
         '''Extract a new index given an iloc key
         '''
+        if self._recache:
+            self._update_array_cache()
+
         if isinstance(key, INT_TYPES):
             # return a tuple if selecting a single row
             # NOTE: if extracting a single row, should be able to get it from IndexLevel without forcing a complete recache
             # NOTE: Selecting a single row may force type coercion before values are added to the tuple; i.e., a datetime64 will go to datetime.date before going to the tuple
             return tuple(self._blocks._extract_array(row_key=key)) #type: ignore
 
+        index_constructors = tuple(self._levels.index_types())
         tb = self._blocks._extract(row_key=key)
 
         return self.__class__._from_type_blocks(tb,
                 name=self._name,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 own_blocks=True,
                 )
 
@@ -1468,7 +1246,6 @@ class IndexHierarchy(IndexBase):
         return self._extract_iloc(key)
 
     #---------------------------------------------------------------------------
-
     def _extract_getitem_astype(self, key: GetItemKeyType) -> 'IndexHierarchyAsType':
         '''Given an iloc key (using integer positions for columns) return a configured IndexHierarchyAsType instance.
         '''
@@ -1485,6 +1262,9 @@ class IndexHierarchy(IndexBase):
             ) -> np.ndarray:
         '''Always return an NP array.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         values = self._blocks.values
         array = operator(values)
         array.flags.writeable = False
@@ -1502,6 +1282,9 @@ class IndexHierarchy(IndexBase):
         from static_frame.core.series import Series
         from static_frame.core.frame import Frame
 
+        if self._recache:
+            self._update_array_cache()
+
         if isinstance(other, (Series, Frame)):
             raise ValueError('cannot use labelled container as an operand.')
 
@@ -1513,6 +1296,8 @@ class IndexHierarchy(IndexBase):
         if isinstance(other, Index):
             other = other.values
         elif isinstance(other, IndexHierarchy):
+            if other._recache:
+                other._update_array_cache()
             other = other._blocks
 
         tb = self._blocks._ufunc_binary_operator(
@@ -1535,6 +1320,9 @@ class IndexHierarchy(IndexBase):
         Returns:
             immutable NumPy array.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         dtype = None if not dtypes else dtypes[0] # must be a tuple
         values = self._blocks.values
 
@@ -1556,33 +1344,25 @@ class IndexHierarchy(IndexBase):
     def __iter__(self) -> tp.Iterator[tp.Tuple[tp.Hashable, ...]]:
         '''Iterate over labels.
         '''
-        # Don't use .values, as that can coerce types
-        yield from zip(*map(self.values_at_depth, range(self.depth)))
+        # NOTE: by iterating from levels, we avoid type casting to a row
+        yield from self._levels.__iter__()
 
     def __reversed__(self) -> tp.Iterator[tp.Tuple[tp.Hashable, ...]]:
         '''
         Returns a reverse iterator on the index labels.
         '''
+        if self._recache:
+            self._update_array_cache()
         for array in self._blocks.axis_values(1, reverse=True):
             yield tuple(array)
 
-    def __contains__(self, value: tp.Tuple[tp.Hashable]) -> bool: # type: ignore
+    def __contains__(self, #type: ignore
+            value: tp.Tuple[tp.Hashable]
+            ) -> bool:
         '''Determine if a leaf loc is contained in this Index.
         '''
-        # TODO: Can this be optimized, or is all the optimization already done in _loc_to_iloc?
-        try:
-            result = self._loc_to_iloc(value)
-        except KeyError:
-            return False
-
-        if isinstance(result, np.ndarray):
-            return bool(result.size)
-
-        if isinstance(result, list):
-            return bool(result)
-
-        return True
-
+        # levels only, no need to recache as this is what has been mutated
+        return self._levels.__contains__(value)
     #---------------------------------------------------------------------------
     # utility functions
 
@@ -1606,10 +1386,14 @@ class IndexHierarchy(IndexBase):
         else: # is an int
             pos = depth_level
 
-        if pos is not None: # i.e. a single level
-            return self._indices[pos].values
-
-        return ufunc_unique(array2d_to_array1d(self.values_at_depth(sel)))
+        if pos is not None:
+            if pos == 0:
+                return self._levels.index.values
+            return np.unique(
+                    concat_resolved(
+                    list(self._levels.index_array_at_depth(pos))
+                    ))
+        return np.unique(array2d_to_array1d(self.values_at_depth(sel)))
 
     @doc_inject()
     def equals(self,
@@ -1629,36 +1413,27 @@ class IndexHierarchy(IndexBase):
             {compare_class}
             {skipna}
         '''
-        # NOTE: do not need to udpate array cache, as can compare elements in levels
+        # NOTE: do not need to udpate array cache, as can compare elemetns in levels
         if id(other) == id(self):
             return True
 
         if compare_class and self.__class__ != other.__class__:
             return False
-
-        if not isinstance(other, IndexHierarchy):
+        elif not isinstance(other, IndexHierarchy):
             return False
 
         # same type from here
         if self.shape != other.shape:
             return False
-
         if compare_name and self.name != other.name:
             return False
 
-        if compare_dtype and not self.dtypes.equals(other.dtypes):
-            return False
-
-        if compare_class:
-            for self_index, other_index in zip(self._indices, other._indices):
-                if self_index.__class__ != other_index.__class__:
-                    return False
-
-        for i in range(self.depth):
-            if not arrays_equal(self.values_at_depth(i), other.values_at_depth(i), skipna=skipna):
-                return False
-
-        return True
+        return self._levels.equals(other._levels, #type: ignore
+                compare_name=compare_name,
+                compare_dtype=compare_dtype,
+                compare_class=compare_class,
+                skipna=skipna,
+                )
 
     @doc_inject(selector='sort')
     def sort(self: IH,
@@ -1674,12 +1449,16 @@ class IndexHierarchy(IndexBase):
             {kind}
             {key}
         '''
+        if self._recache:
+            self._update_array_cache()
+
         order = sort_index_for_order(self, kind=kind, ascending=ascending, key=key) #type: ignore [arg-type]
 
         blocks = self._blocks._extract(row_key=order)
+        index_constructors = tuple(self._levels.index_types())
 
         return self.__class__._from_type_blocks(blocks,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 name=self._name,
                 own_blocks=True
                 )
@@ -1688,6 +1467,9 @@ class IndexHierarchy(IndexBase):
         '''
         Return a Boolean array showing True where one or more of the passed in iterable of labels is found in the index.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         matches = []
         for seq in other:
             if not hasattr(seq, '__iter__'):
@@ -1706,12 +1488,16 @@ class IndexHierarchy(IndexBase):
     def roll(self, shift: int) -> 'IndexHierarchy':
         '''Return an :obj:`IndexHierarchy` with values rotated forward and wrapped around (with a positive shift) or backward and wrapped around (with a negative shift).
         '''
+        if self._recache:
+            self._update_array_cache()
+
         blocks = TypeBlocks.from_blocks(
                 self._blocks._shift_blocks(row_shift=shift, wrap=True)
                 )
+        index_constructors = tuple(self._levels.index_types())
 
         return self.__class__._from_type_blocks(blocks,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 name=self._name,
                 own_blocks=True
                 )
@@ -1723,10 +1509,14 @@ class IndexHierarchy(IndexBase):
         Args:
             {value}
         '''
+        if self._recache:
+            self._update_array_cache()
+
         blocks = self._blocks.fill_missing_by_unit(value, None, func=isna_array)
+        index_constructors = tuple(self._levels.index_types())
 
         return self.__class__._from_type_blocks(blocks,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 name=self._name,
                 own_blocks=True
                 )
@@ -1736,17 +1526,22 @@ class IndexHierarchy(IndexBase):
             *,
             seed: tp.Optional[int] = None,
             ) -> tp.Tuple['IndexHierarchy', np.ndarray]:
+        if self._recache:
+            self._update_array_cache()
 
         # sort to ensure hierarchability
         key = array_sample(self.positions, count=count, seed=seed, sort=True)
         blocks = self._blocks._extract(row_key=key)
 
+        index_constructors = tuple(self._levels.index_types())
+
         container = self.__class__._from_type_blocks(blocks,
-                index_constructors=self._index_constructors,
+                index_constructors=index_constructors,
                 name=self._name,
                 own_blocks=True
                 )
         return container, key
+
 
     @doc_inject(selector='searchsorted', label_type='iloc (integer)')
     def iloc_searchsorted(self,
@@ -1771,7 +1566,7 @@ class IndexHierarchy(IndexBase):
             raise NotImplementedError('A single label (as a tuple) or multiple labels (as a list) must be provided.')
 
         dt_pos = np.array([issubclass(idx_type, IndexDatetime)
-                for idx_type in self._index_constructors])
+                for idx_type in self._levels.index_types()])
         has_dt = dt_pos.any()
 
         values_for_match = np.empty(len(match_pre), dtype=object)
@@ -1786,6 +1581,7 @@ class IndexHierarchy(IndexBase):
         if is_element:
             return post[0] #type: ignore [no-any-return]
         return post #type: ignore [no-any-return]
+
 
     @doc_inject(selector='searchsorted', label_type='loc (label)')
     def loc_searchsorted(self,
@@ -1821,12 +1617,18 @@ class IndexHierarchy(IndexBase):
         post.flags.writeable = False
         return post #type: ignore [no-any-return]
 
+
+
+
     #---------------------------------------------------------------------------
     # export
 
     def _to_frame(self,
             constructor: tp.Type['Frame']
             ) -> 'Frame':
+
+        if self._recache:
+            self._update_array_cache()
 
         return constructor(
                 self._blocks.copy(),
@@ -1854,39 +1656,23 @@ class IndexHierarchy(IndexBase):
         '''
         import pandas
 
+        if self._recache:
+            self._update_array_cache()
+
         # must copy to get a mutable array
-        mi = pandas.MultiIndex(
-                levels=[index.values.copy() for index in self._indices],
-                codes=[arr.copy() for arr in self._indexers],
-                )
+        arrays = tuple(a.copy() for a in self._blocks.axis_values(axis=0))
+        mi = pandas.MultiIndex.from_arrays(arrays)
+
         mi.name = self._name
         mi.names = self.names
         return mi
 
-    def _build_tree_at_depth_from_mask(self, depth: int, mask: np.ndarray) -> tp.Union[TreeNodeT, Index]:
-
-        if depth == self.depth - 1:
-            values = self._indices[depth][self._indexers[depth][mask]]
-            return self._indices[depth].__class__(values)
-
-        tree: TreeNodeT = {}
-
-        index_at_depth = self._indices[depth]
-        indexer_at_depth = self._indexers[depth]
-
-        for i in ufunc_unique(indexer_at_depth[mask]):
-            tree[index_at_depth[i]] = self._build_tree_at_depth_from_mask(depth + 1, mask & (indexer_at_depth == i))
-
-        return tree
-
     def to_tree(self) -> TreeNodeT:
         '''Returns the tree representation of an IndexHierarchy
         '''
-        tree = self._build_tree_at_depth_from_mask(0, np.ones(len(self), dtype=bool))
-        assert isinstance(tree, dict) # mypy
-        return tree
+        return self._levels.traverse() # type: ignore
 
-    def flat(self) -> Index:
+    def flat(self) -> IndexBase:
         '''Return a flat, one-dimensional index of tuples for each level.
         '''
         return self._INDEX_CONSTRUCTOR(self.__iter__(), name=self._name)
@@ -1898,29 +1684,30 @@ class IndexHierarchy(IndexBase):
             ) -> IH:
         '''Return an IndexHierarchy with a new root (outer) level added.
         '''
-        index_cls = self._INDEX_CONSTRUCTOR if index_constructor is None else index_constructor._MUTABLE_CONSTRUCTOR # type: ignore
-
-        if self.STATIC:
-            indices = [index_cls((level,)), *self._indices]
+        if self.STATIC: # can reuse levels
+            levels_src = self._levels
         else:
-            indices = [index_cls((level,)), *(idx.copy() for idx in self._indices)]
+            levels_src = self._levels.to_index_level()
 
-        # Indexers are always immutable
-        new_indexer = np.full(self.__len__(), 0, dtype=int)
-        new_indexer.flags.writeable = False
-        indexers = [new_indexer, *self._indexers]
+        index_cls = self._INDEX_CONSTRUCTOR if index_constructor is None else index_constructor
 
-        def gen_blocks() -> tp.Iterator[np.ndarray]:
-            yield np.full(self.__len__(), indices[0].values)
-            yield from self._blocks._blocks
+        levels = self._LEVEL_CONSTRUCTOR(
+                index=index_cls((level,)),
+                targets=ArrayGO([levels_src], own_iterable=True),
+                offset=0,
+                own_index=True
+                )
+        # can transfrom TypeBlocks appropriately and pass to constructor
+        if not self._recache: # if we have TypeBlocks
+            array = np.full(self.__len__(), level)
+            array.flags.writeable = False
+            blocks = TypeBlocks.from_blocks(chain((array,), self._blocks._blocks))
+            return self.__class__(levels,
+                    name=self._name,
+                    blocks=blocks,
+                    own_blocks=True)
 
-        return self.__class__(
-            indices=indices,
-            indexers=indexers,
-            name=self.name,
-            _blocks=TypeBlocks.from_blocks(gen_blocks()),
-            _own_blocks=True,
-        )
+        return self.__class__(levels, name=self._name)
 
     def level_drop(self,
             count: int = 1,
@@ -1930,6 +1717,9 @@ class IndexHierarchy(IndexBase):
         Args:
             count: A positive value is the number of depths to remove from the root (outer) side of the hierarchy; a negative value is the number of depths to remove from the leaf (inner) side of the hierarchy.
         '''
+        if self._recache:
+            self._update_array_cache()
+
         # NOTE: this was implement with a bipolar ``count`` to specify what to drop, but it could have been implemented with a depth level specifier, supporting arbitrary removals. The approach taken here is likely faster as we reuse levels.
         if self._name_is_names():
             if count < 0:
@@ -1942,27 +1732,55 @@ class IndexHierarchy(IndexBase):
             name = self._name
 
         if count < 0: # remove from inner
-            if count <= (1 - self.depth):
-                return self._index_constructors[-1](self._blocks.iloc[:,0].values.ravel(), name=name)
+            levels = self._levels.to_index_level()
+            for _ in range(abs(count)):
+                levels_stack = [levels]
+                while levels_stack:
+                    level = levels_stack.pop()
+                    # check to see if children of this target are leaves
+                    if level.targets[0].targets is None: #type: ignore
+                        level.targets = None
+                    else:
+                        levels_stack.extend(level.targets) #type: ignore
+                if levels.targets is None:  # if no targets, at the root
+                    break
 
-            return self.__class__(
-                    indices=self._indices[:count],
-                    indexers=self._indexers[:count],
-                    name=name,
-                    _blocks=self._blocks[:count],
-                    _own_blocks=self.STATIC,
-                    )
+            if levels.targets is None: # fall back to 1D index
+                return levels.index.rename(name)
+
+            # if we have TypeBlocks and levels is the same length
+            if levels.__len__() == self.__len__():
+                blocks = self._blocks.iloc[NULL_SLICE, :count]
+                return self.__class__(levels,
+                        blocks=blocks,
+                        own_blocks=True,
+                        name=name,
+                        )
+            return self.__class__(levels, name=name)
 
         elif count > 0: # remove from outer
-            if count >= (self.depth - 1):
-                return self._index_constructors[0](self._blocks.iloc[:,-1].values.ravel(), name=name)
+            levels = self._levels.to_index_level()
+            for _ in range(count):
+                targets = []
+                labels = []
+                for target in levels.targets: #type: ignore
+                    labels.extend(target.index)
+                    if target.targets is not None:
+                        targets.extend(target.targets)
+                index = levels.index.__class__(labels)
+                if not targets:
+                    return index.rename(name)
+                levels = levels.__class__(
+                        index=index,
+                        targets=ArrayGO(targets, own_iterable=True))
 
-            return self.__class__(
-                    indices=self._indices[count:],
-                    indexers=self._indexers[count:],
+            # lengths must be equal if innner index components are unique; otherwise exception will have been raised
+            # assert levels.__len__() == self.__len__()
+            blocks = self._blocks.iloc[NULL_SLICE, count:]
+            return self.__class__(levels,
                     name=name,
-                    _blocks=self._blocks.iloc[:,count:],
-                    _own_blocks=self.STATIC,
+                    blocks=blocks,
+                    own_blocks=True,
                     )
 
         raise NotImplementedError('no handling for a 0 count drop level.')
@@ -1976,123 +1794,46 @@ class IndexHierarchyGO(IndexHierarchy):
     STATIC = False
 
     _IMMUTABLE_CONSTRUCTOR = IndexHierarchy
+    _LEVEL_CONSTRUCTOR = IndexLevelGO
     _INDEX_CONSTRUCTOR = IndexGO
 
     __slots__ = (
-            '_name',
-            '_indices',
-            '_indexers',
+            '_levels', # IndexLevel
             '_blocks',
-            '_values',
+            '_recache',
+            '_name'
             )
 
-    _indices: tp.List[IndexGO] # type: ignore
+    _levels: IndexLevelGO
 
     def append(self, value: tp.Sequence[tp.Hashable]) -> None:
         '''
         Append a single label to this index.
         '''
-        if value in self: # type: ignore
-            raise ErrorInitIndexNonUnique(f"The label '{value}' is already in the index.")
-
-        label_indexers = []
-
-        for depth, label_at_depth in enumerate(value):
-            if label_at_depth in self._indices[depth]:
-                label_index = self._indices[depth]._loc_to_iloc(label_at_depth)
-            else:
-                label_index = len(self._indices[depth])
-                self._indices[depth].append(label_at_depth)
-
-            label_indexers.append(label_index)
-
-        for i, label_index in enumerate(label_indexers):
-            self._indexers[i] = np.append(self._indexers[i], label_index)
-            self._indexers[i].flags.writeable = False
-
-        # No need to ensure uniqueness! It's already been checked.
-        self._blocks = self._gen_blocks_from_self()
-        self._values = self._blocks.values
+        self._levels.append(value)
+        self._recache = True
 
     def extend(self, other: IndexHierarchy) -> None:
         '''
         Extend this IndexHiearchy in-place
         '''
-        for depth, (self_index, other_index) in enumerate(zip(self._indices, other._indices)):
-
-            intersection = self_index.intersection(other_index)
-            if not intersection.size:
-                del intersection
-
-                starting_len = len(self_index)
-
-                # Easy case! We can simply append
-                self_index.extend(other_index)
-
-                new_indexer = np.hstack((self._indexers[depth], other._indexers[depth] + starting_len))
-                new_indexer.flags.writeable = False
-                self._indexers[depth] = new_indexer
-                continue
-
-            if len(intersection) == len(self_index) == len(other_index):
-                del intersection
-
-                if self_index.equals(other_index):
-                    # Easy case! We just have to append the indexers; no change needed to the index
-
-                    new_indexer = np.hstack((self._indexers[depth], other._indexers[depth]))
-                    new_indexer.flags.writeable = False
-                    self._indexers[depth] = new_indexer
-                    continue
-
-                # Same labels, but different order. We have to remap the indexers.
-                indexer_remap = other_index.iter_label().apply(self_index._loc_to_iloc)
-
-                remap_indexer = indexer_remap[other._indexers[depth]]
-                new_indexer = np.hstack((self._indexers[depth], remap_indexer))
-                new_indexer.flags.writeable = False
-                self._indexers[depth] = new_indexer
-                continue
-
-            starting_len = len(self_index)
-
-            self_index.extend(other_index[~other_index.isin(intersection)])
-
-            def remap(k: tp.Hashable) -> int:
-                if k in intersection:
-                    return self_index._loc_to_iloc(k) # type: ignore
-                return -1
-
-            offset = starting_len - len(intersection)
-            indexer_remap = other_index.iter_label().apply(remap)
-            del intersection
-
-            remap_indexer = indexer_remap[other._indexers[depth]]
-
-            mask = remap_indexer == -1
-
-            remap_indexer[mask] = (other._indexers[depth][mask] + offset)
-            new_indexer = np.hstack((self._indexers[depth], remap_indexer))
-            new_indexer.flags.writeable = False
-            self._indexers[depth] = new_indexer
-
-        # No need to ensure uniqueness! It's already been checked.
-        self._blocks = self._gen_blocks_from_self()
-        self._values = self._blocks.values
-        self._ensure_uniqueness(self._values)
+        self._levels.extend(other._levels)
+        self._recache = True
 
     def __copy__(self: IH) -> IH:
         '''
         Return a shallow copy of this IndexHierarchy.
         '''
-        return self.__class__(
-                indices=[index.copy() for index in self._indices],
-                indexers=self._indexers,
-                name=self._name,
-                _blocks=self._blocks.copy(),
-                _own_blocks=True,
-                )
+        if self._recache:
+            self._update_array_cache()
 
+        blocks = self._blocks.copy()
+        return self.__class__(
+                levels=self._levels.to_index_level(), # recursive deep copy
+                name=self._name,
+                blocks=blocks,
+                own_blocks=True,
+                )
 
 # update class attr on Index after class initialization
 IndexHierarchy._MUTABLE_CONSTRUCTOR = IndexHierarchyGO
@@ -2113,6 +1854,9 @@ class IndexHierarchyAsType:
 
         from static_frame.core.index_datetime import dtype_to_index_cls
         container = self.container
+
+        if container._recache:
+            container._update_array_cache()
 
         # use TypeBlocks in both situations to avoid double casting
         blocks = TypeBlocks.from_blocks(
@@ -2137,3 +1881,4 @@ class IndexHierarchyAsType:
                 index_constructors=index_constructors,
                 own_blocks=True
                 )
+
