+++ static_frame/core/pivot.py
@@ -546,7 +546,7 @@ def pivot_derive_constructors(*,
     if expand_src.depth == 1:
         expand_types = [expand_src.__class__]
     else:
-        expand_types = expand_src.index_types.values.tolist()
+        expand_types = list(expand_src._levels.index_types()) #type: ignore
 
     if contract_src.depth == 1:
         expand_types.append(contract_src.__class__)
