+++ static_frame/core/quilt.py
@@ -898,12 +898,7 @@ class Quilt(ContainerBase, StoreClientMixin):
         if isinstance(axis_map_sub, tuple): # type: ignore
             bus_keys = (axis_map_sub[0],) #type: ignore
         else:
-            # get the outer level, or just the unique frame labels needed
-            labels_at_depth = axis_map_sub.values_at_depth(0)
-            label_indexes = sorted(np.unique(labels_at_depth, return_index=True)[1])
-            bus_keys = labels_at_depth[label_indexes]
-            del labels_at_depth
-            del label_indexes
+            bus_keys = axis_map_sub._levels.index
 
         for key_count, key in enumerate(bus_keys):
             sel_component = sel[self._axis_hierarchy._loc_to_iloc(HLoc[key])]
@@ -984,15 +979,11 @@ class Quilt(ContainerBase, StoreClientMixin):
 
         # get ordered unique Bus labels
         axis_map_sub = self._axis_hierarchy.iloc[sel_key]
-        if isinstance(axis_map_sub, tuple): # type: ignore
+        if isinstance(axis_map_sub, tuple): #type: ignore
             frame_labels = (axis_map_sub[0],) #type: ignore
         else:
             # get the outer level, or just the unique frame labels needed
-            labels_at_depth = axis_map_sub.values_at_depth(0)
-            label_indexes = sorted(np.unique(labels_at_depth, return_index=True)[1])
-            frame_labels = labels_at_depth[label_indexes]
-            del labels_at_depth
-            del label_indexes
+            frame_labels = axis_map_sub._levels.index
 
         for key_count, key in enumerate(frame_labels):
             # get Boolean segment for this Frame
