+++ static_frame/core/util.py
@@ -1,7 +1,6 @@
 from collections import abc
 from collections import defaultdict
 from collections import namedtuple
-from collections import Counter
 from enum import Enum
 from functools import partial
 from functools import reduce
@@ -881,8 +880,6 @@ def ufunc_unique(
         array: np.ndarray,
         *,
         axis: tp.Optional[int] = None,
-        return_inverse: bool = False,
-        return_counts: bool = False,
         ) -> np.ndarray:
     '''
     Extended functionality of the np.unique ufunc, to handle cases of mixed typed objects, where NP will fail in finding unique values for a heterogenous object type.
@@ -893,7 +890,7 @@ def ufunc_unique(
     if array.dtype.kind == 'O':
         if axis is None or array.ndim < 2:
             try:
-                return np.unique(array, return_inverse=return_inverse, return_counts=return_counts)
+                return np.unique(array)
             except TypeError: # if unorderable types
                 # np.unique will give TypeError: The axis argument to unique is not supported for dtype object
                 pass
@@ -909,55 +906,14 @@ def ufunc_unique(
             else:
                 array_iter = array2d_to_tuples(array.T)
 
-        if return_inverse:
-            inverse = np.empty(len(array_iter), dtype=np.intp)
-
         # Use a dict to retain order; this will break for non hashables
-        if return_counts:
-            if not return_inverse:
-                store: tp.Mapping[tp.Hashable, tp.Union[int, None]] = Counter(array_iter)
-            else:
-                store = Counter()
-                indices: tp.Dict[tp.Hashable, int] = {}
-                for i, element in enumerate(array_iter):
-                    store[element] += 1
-                    indices.setdefault(element, len(indices))
-                    inverse[i] = indices[element]
-
-                indices.clear()
-
-            counts = np.empty(len(store), dtype=np.intp)
-            counts[:] = tuple(store.values())
-        else:
-            if not return_inverse:
-                store = dict.fromkeys(array_iter)
-            else:
-                store = {}
-                indices = {}
-                for i, element in enumerate(array_iter):
-                    store[element] = None
-                    indices.setdefault(element, len(indices))
-                    inverse[i] = indices[element]
-
-                indices.clear()
-
+        store = dict.fromkeys(array_iter)
         array = np.empty(len(store), dtype=object)
         array[:] = tuple(store)
-
-        if return_inverse and return_counts:
-            return array, inverse, counts
-
-        if return_inverse:
-            return array, inverse
-
-        if return_counts:
-            return array, counts
-
         return array
 
-
     # all other types, use the main ufunc
-    return np.unique(array, axis=axis, return_inverse=return_inverse, return_counts=return_counts)
+    return np.unique(array, axis=axis)
 
 
 def roll_1d(array: np.ndarray,
@@ -2082,10 +2038,6 @@ def _ufunc_set_1d(
 
     if is_union:
         post = func(array, other)
-
-        # Preserve order when union is the same as array!
-        # if len(post) == len(array) == len(other):
-        #     post = array.copy()
     else:
         post = func(array, other, assume_unique=assume_unique) #type: ignore
